[PR c++/85437] accept static_casted ptrmem in constexpr
A static_cast of a pointer to data member used to wrap the PTRMEM_CST
in a NOP_EXPR, but the NOP_EXPR is taken, in constexpr, as evidence
that there was a reinterpret_cast in the expression.  While
reinterpret_casts are to be rejected in constexprs, static_casts are
ok.

Thus, avoid introducing the NOP_EXPR in static_casts, folding the
converted-to type into the PTRMEM_CST type.

This requires PTRMEM_CST constant expansion to deal with such up and
downcasts.


for  gcc/cp/ChangeLog

	PR c++/85437
	* expr.c (cplus_expand_constant): Compute deltas for up and
	downcasts.
	* type.c (convert_ptrmem): Convert ptrmem type for static
	cast.

for  gcc/testsuite/ChangeLog

	PR c++/85437
	* g++.dg/cpp0x/pr85437.C: New.
	* g++.dg/cpp0x/pr85437-2.C: New.
	* g++.dg/cpp0x/pr85437-3.C: New.